Rename mock files `mock[type]` for better sorting
