+++ src/components/skills/mockSkills.ts
